refactor(dashboard) : split pages and made component for reusable and add intial pages
