+++ src/pages/dashboard/threads/ManageThread.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ManageThread = () => {
+  return (
+    <div className="site-layout-background">
+      <h1>Cek</h1>
+    </div>
+  );
+};
+
+export default ManageThread;
