+++ src/pages/dashboard/users/ManageUsers.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ManageUsers = () => {
+  return <div className="site-layout-background">Manage Users</div>;
+};
+
+export default ManageUsers;
